fix(watch): Int64 en measuredAtMs — Int de 32 bits crasheaba el reloj
EL CRASH: la app del reloj moria con EXC_BREAKPOINT apenas empezaba un entreno.
Diagnosticado sobre el reporte de crash real del dispositivo (build 19, watchOS
26.6), simbolizado con el dSYM del archive — UUID verificado contra el binario
que crasheo:

    EffortSnapshot.context(measuredAt:)   EffortBroadcastRules.swift:35  <- TRAP
    EffortRelay.publish(bpm:kcal:...)     EffortRelay.swift:61
    TreinoWatch_Watch_AppApp.publicarEsfuerzo()
    closure en el body de la App          (el .onChange de heartRate/activeEnergy)

LA CAUSA: `Int(measuredAt.timeIntervalSince1970 * 1000)`.

watchOS corre arm64_32 — verificado con `lipo -archs` sobre el binario del IPA.
Ahi `Int` es de 32 bits. Los milisegundos desde 1970 son ~1,787e12 y el maximo
de Int32 es 2,147e9: lo excede por un factor de 832. Y `Int(Double)` no trunca
ni devuelve basura cuando el valor no entra — TRAPEA. El proceso muere en el
primer dato de pulso que llega.

En el iPhone `Int` son 64 bits y el mismo codigo es correcto. Por eso sobrevivio:
es letal solo de un lado del emparejamiento.

POR QUE APARECIO RECIEN AHORA, que es la parte interesante:
el bug estaba desde que se escribio el relay, pero era INALCANZABLE. Sin el
indice compuesto de `sessions` la pantalla HOY fallaba, y el boton "Empezar"
vive adentro de esa pantalla — asi que el reloj nunca entraba en modo entreno,
HealthKit nunca arrancaba, y `publicarEsfuerzo()` no corria jamas. Arreglar el
indice hizo fluir pulso y calorias por primera vez, y con eso el trap se volvio
alcanzable. Un arreglo destapo un crash dormido.

POR QUE NINGUN GATE LO AGARRO: `scripts/test_watch_swift.sh` compila con
`swiftc` para el HOST, o sea arm64 de 64 bits, donde la conversion es valida.
El harness es estructuralmente incapaz de reproducir un overflow de 32 bits.

El test que se agrega verifica el TIPO, que es lo unico observable desde el
host, y alcanza: `Int64` es justamente lo que evita el trap en el reloj.
Verificado en ROJO — con `Int` fallan 2 chequeos ("vino nil" al castear a
Int64). 114 chequeos en verde con el fix.

Contrato con Dart intacto: `WatchEffort.tryParse` pide `ms is int`, y los int de
Dart son de 64 bits.

Se reviso todo el target: es la UNICA conversion de este tipo.

+++ ios/TreinoWatch Watch App/EffortBroadcastRules.swift
@@ -30,9 +30,24 @@ struct EffortSnapshot: Equatable {
     /// Cambiar una acá sin cambiarla alla deja al telefono sin datos y sin
     /// ningun error visible.
     func context(measuredAt: Date) -> [String: Any] {
+        // `Int64` y NO `Int`. Esto CRASHEABA el reloj.
+        //
+        // watchOS corre arm64_32: los punteros son de 32 bits y `Int` tambien.
+        // Los milisegundos desde 1970 rondan 1,79e12, que excede el maximo de
+        // Int32 (2,15e9) por un factor de ~832. Y `Int(Double)` no trunca ni
+        // devuelve basura cuando no entra: TRAPEA. El proceso muere con
+        // EXC_BREAKPOINT en el primer dato de pulso que llega.
+        //
+        // En el iPhone `Int` son 64 bits y esto anda perfecto, que es
+        // exactamente por que sobrevivio tanto: el mismo codigo es correcto de
+        // un lado del emparejamiento y letal del otro.
+        //
+        // Los tests puros del reloj TAMPOCO lo agarraban: corren en el host con
+        // `swiftc`, o sea arm64 de 64 bits. El test que se agrego abajo verifica
+        // el TIPO, que es lo unico observable desde el host.
         var payload: [String: Any] = [
             "kind": "watchEffort",
-            "measuredAtMs": Int(measuredAt.timeIntervalSince1970 * 1000),
+            "measuredAtMs": Int64(measuredAt.timeIntervalSince1970 * 1000),
         ]
         if let bpm { payload["bpm"] = bpm }
         if let kcal { payload["kcal"] = kcal }
